requirements.txt
behave==1.2.6
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
commandbus==1.0.1
Flask==1.0.2
Flask-WTF==0.14.2
google-api-core==1.1.2
google-auth==1.4.1
google-cloud-bigquery==1.1.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.13
gunicorn==19.8.1
idna==2.6
iniherit==0.3.9
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
parse==1.8.2
parse-type==0.4.2
passlib==1.7.1
protobuf==3.5.2.post1
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-slugify==1.2.5
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
yoyo-migrations==5.0.5
